+++ CHANGELOG.md
@@ -5,6 +5,8 @@ which adheres to [semantic versioning](https://semver.org).
 
 ## unreleased
 
+* Add new configuration file and `hashless` option (file glob array) to upload
+  files to the CDN without their md5 hash appended.
 * Add upload support for PDF files.
 
 ## v2.0.1 2019-07-30